% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_meas.R
\name{acc_meas}
\alias{acc_meas}
\title{Accuracy measures}
\usage{
acc_meas(tab)
}
\arguments{
\item{tab}{confusion matrix with prediction as columns}
}
\value{
A list containing the following entries
\itemize{
   \item{True positive rate}
   \item{True negative rate}
   \item{Area under the curve}
   \item{ggplot object for the ROC curve}
 }
}
\description{
Provides measures to evaluate the performance of the classifier.
}
