% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Euclidean distance}
\usage{
distance(data, y)
}
\arguments{
\item{data}{the complete set of features (both classes)}

\item{y}{outcome variable to be classified}
}
\value{
a vector containing the distances
}
\description{
Calculate the euclidean distance of the features in \code{data}
of each observation from the center of the majority class
}
