% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_likelihood.R
\name{est_likelihood}
\alias{est_likelihood}
\title{Estimate log-likelihood}
\usage{
est_likelihood(data, y)
}
\arguments{
\item{data}{the complete set of features (both classes)}

\item{y}{outcome variable to be classified}
}
\value{
a vector containing the log-likelihood estimated
}
\description{
Estimate the log-likelihood of each observation originating from the joint distribution of the features in \code{data} for the majority class
}
